Update the koans, in this session experiment with hashes and symbols
In this session, i made hashes, strings, symbols, objects and nils.
I leanend that in ruby all elemnts are objects, for this reason, ruby
can convert easily from symbols to strings and other element.

Hashes are like a asociative arrays, but not are arrays, and nil are too
a object.